fix(service): update SignFileService to use IPkcs12SignEngineHandler
